backend/aws/service/functions/requirements.txt
boto3
nose2
pillow
requests
matplotlib
google-cloud-firestore